+++ src/main/resources/db/migrations/changes/18112019_create_table_week.sql
@@ -1,8 +1,8 @@
 create table week
 (
-    id           bigint primary key not null,
+    id           serial primary key not null,
     week_day     varchar(100),
-    lesson_index bigint,
+    lesson_index serial,
     subject_name varchar(100),
     week_type    varchar(100)
 )
